security: secure the endpoints and protected the page

+++ app/api/issues/[id]/route.ts
@@ -1,11 +1,17 @@
 import prisma from "@/prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 import { IssueSchema } from "@/app/ValidationSchemas";
+import authOptions from "@/app/auth/AuthOptions";
+import { getServerSession } from "next-auth";
 
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await  getServerSession(authOptions);
+  if(!session)
+    return NextResponse.json({},{status:401})
+  
   const body = await request.json();
   const validation = IssueSchema.safeParse(body);
   if (!validation.success)
@@ -29,6 +35,10 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await  getServerSession(authOptions);
+  if(!session)
+    return NextResponse.json({},{status:401})
+   
   const issue = await prisma.issue.findUnique({ where: { id: parseInt(params.id) } });
   if(!issue) return NextResponse.json({error: "Invalid Issue"}, {status: 404});
   await prisma.issue.delete({where: {id: issue.id}});
